Synchronize Branches

+++ admin_page/app/Http/Controllers/OrderController.php
@@ -10,6 +10,9 @@ use App\Client;
 use App\Exports\OrdersExport;
 use App\Product;
 use App\TimeBlock;
+use App\ProductFormat;
+use Illuminate\Support\Carbon;
+use Illuminate\Http\Request;
 
 class OrderController extends Controller
 {
@@ -20,15 +23,29 @@ class OrderController extends Controller
      */
     public function index(FilterRequest $request)
     {
-        $client_type = $request->client_type;
-        $time_interval_start = $request->time_interval_start;
-        $time_interval_end = $request->time_interval_end;
-        $town_id = $request->town_id;
-        $order_status = $request->order_status;
+        $this->authorize('view', Order::class);
+
+        $client_type = $request->query('client_type');
+        $time_interval_start = $request->query('time_interval_start');
+        $time_interval_end = $request->query('time_interval_end');
+        $town_id = $request->query('town_id');
+        $time_block_id = $request->query('time_block_id');
+        $delivery_status = $request->query('delivery_status');
+        $payment_status = $request->query('payment_status');
+
+        session(['client_type' => $client_type]);
+        session(['time_interval_start' => $time_interval_start]);
+        session(['time_interval_end' => $time_interval_end]);
+        session(['town_id' => $town_id]);
+        session(['time_block_id' => $time_block_id]);
+        session(['delivery_status' => $delivery_status]);
+        session(['payment_status' => $payment_status]);
 
         $orders = Order::search($request);
         $towns = Town::pluck('name', 'id');
-        $orderStatuses = Order::getStatuses();
+        $timeBlocks = TimeBlock::orderBy('start')->get();
+        $deliveryStatuses = Order::getDeliveryStatuses();
+        $paymentStatuses = Order::getPaymentStatuses();
         $clientTypes = Client::getClientTypes();
 
         if($request->has('generate_excel') && $request->generate_excel == true) {
@@ -36,7 +53,7 @@ class OrderController extends Controller
         }
         $orders = $orders->paginate(Order::ITEMS_PER_PAGE);
 
-        return view('orders.index', compact('orders', 'towns', 'orderStatuses', 'clientTypes', 'client_type', 'time_interval_start', 'time_interval_end', 'town_id', 'order_status'))
+        return view('orders.index', compact('orders', 'towns', 'timeBlocks', 'deliveryStatuses', 'paymentStatuses', 'clientTypes', 'client_type', 'time_interval_start', 'time_interval_end', 'town_id', 'time_block_id', 'delivery_status', 'payment_status'))
             ->with('rowItem', $this->rowNumber(request()->input('page', 1), Order::ITEMS_PER_PAGE) );
     }
 
@@ -47,12 +64,16 @@ class OrderController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Order::class);
+
         $towns = Town::pluck('name', 'id');
         $products = Product::pluck('name', 'id');
-        $orderStatuses = Order::getStatuses();
+        $formats = ProductFormat::where('product_id', Product::compounded()->first()->id)->pluck('name', 'id');
+        $deliveryStatuses = Order::getDeliveryStatuses();
+        $paymentStatuses = Order::getPaymentStatuses();
         $timeBlocks = TimeBlock::orderBy('start')->get();
 
-        return view('orders.create', compact('towns', 'products', 'orderStatuses', 'timeBlocks'));
+        return view('orders.create', compact('towns', 'products', 'deliveryStatuses', 'paymentStatuses', 'timeBlocks', 'formats'));
     }
 
     /**
@@ -63,6 +84,8 @@ class OrderController extends Controller
      */
     public function store(OrderRequest $request)
     {
+        $this->authorize('create', Order::class);
+
         Order::createFromForm($request->validated());
 
         return redirect()->route('orders.index')
@@ -77,6 +100,8 @@ class OrderController extends Controller
      */
     public function show(Order $order)
     {
+        $this->authorize('view', Order::class);
+
         return view('orders.show', compact('order'));
     }
 
@@ -88,6 +113,8 @@ class OrderController extends Controller
      */
     public function edit(Order $order)
     {
+        $this->authorize('update', Order::class);
+
         return view('orders.edit', compact('order'));
     }
 
@@ -100,6 +127,8 @@ class OrderController extends Controller
      */
     public function update(OrderRequest $request, Order $order)
     {
+        $this->authorize('update', Order::class);
+
         $order->update($request->validated());
 
         return redirect()->route('orders.index')
@@ -114,17 +143,85 @@ class OrderController extends Controller
      */
     public function destroy(Order $order)
     {
+        $this->authorize('delete', Order::class);
+
         $order->delete();
 
-        return redirect()->route('orders.index')
-            ->with('success', 'Pedido eliminado exitosamente');
+        return redirect()->route('orders.index', [
+                'client_type' => session('client_type'),
+                'time_interval_start' => session('time_interval_start'),
+                'time_interval_end' => session('time_interval_end'),
+                'town_id' => session('town_id'),
+                'time_block_id' => session('time_block_id'),
+                'delivery_status' => session('delivery_status'),
+                'payment_status' => session('payment_status'),
+            ])->with('success', 'Pedido eliminado exitosamente');
     }
 
+    /**
+     * Change order delivery_status to delivered.
+     *
+     * @param  \App\Order  $order
+     * @return \Illuminate\Http\Response
+     */
     public function delivered(Order $order)
     {
+        $this->authorize('deliver', Order::class);
+
         $order->delivered();
 
-        return redirect()->route('orders.index')
-            ->with('success', 'Estado cambiado exitosamente');
+        return redirect()->route('orders.index', [
+                'client_type' => session('client_type'),
+                'time_interval_start' => session('time_interval_start'),
+                'time_interval_end' => session('time_interval_end'),
+                'town_id' => session('town_id'),
+                'time_block_id' => session('time_block_id'),
+                'delivery_status' => session('delivery_status'),
+                'payment_status' => session('payment_status'),
+            ])->with('success', 'Estado cambiado exitosamente');
+    }
+
+    /**
+     * Change order payment_status to paid.
+     *
+     * @param  \App\Order  $order
+     * @return \Illuminate\Http\Response
+     */
+    public function paid(Order $order)
+    {
+        $this->authorize('payment', Order::class);
+
+        $order->paid();
+
+        return redirect()->route('orders.index', [
+                'client_type' => session('client_type'),
+                'time_interval_start' => session('time_interval_start'),
+                'time_interval_end' => session('time_interval_end'),
+                'town_id' => session('town_id'),
+                'time_block_id' => session('time_block_id'),
+                'delivery_status' => session('delivery_status'),
+                'payment_status' => session('payment_status'),
+            ])->with('success', 'Estado cambiado exitosamente');
+    }
+
+    // Functions used in ajax calls to make the order creation dynamic
+    /*
+     * Get the formats of the related product.
+     */
+    public function productFormats(Request $request) {
+        if(is_numeric((int)$request->query('id'))) {
+            return ProductFormat::where('product_id', $request->id)->pluck('name', 'id');
+        }
+        return [];
+    }
+
+    /*
+     * Get the available time blocks of given date.
+     */
+    public function availableTimeBlocks(Request $request) {
+        if($request->has('delivery_date')) {
+            return TimeBlock::getAvailableBlocks(Carbon::parse($request->delivery_date))->map->only(['id', 'start', 'end']);
+        }
+        return [];
     }
 }
