+++ admin_page/app/Http/Controllers/ProductDiscountController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Product;
 use App\ProductDiscount;
 use App\Http\Requests\ProductDiscountRequest;
-use App\Product;
 
 class ProductDiscountController extends Controller
 {
@@ -16,6 +16,8 @@ class ProductDiscountController extends Controller
      */
     public function index(Product $product)
     {
+        $this->authorize('view', ProductDiscount::class);
+
         $productDiscounts = $product->discountsPaginated();
         return view('discounts.index', compact('product', 'productDiscounts'))
             ->with('rowItem', $this->rowNumber(request()->input('page', 1), ProductDiscount::ITEMS_PER_PAGE) );
@@ -29,6 +31,8 @@ class ProductDiscountController extends Controller
      */
     public function create(Product $product)
     {
+        $this->authorize('create', ProductDiscount::class);
+
         return view('discounts.create', compact('product'));
     }
 
@@ -41,6 +45,12 @@ class ProductDiscountController extends Controller
      */
     public function store(ProductDiscountRequest $request, Product $product)
     {
+        $this->authorize('create', ProductDiscount::class);
+
+        if (ProductDiscount::where('product_id', $product->id)->intersecting($request->min_quantity, $request->max_quantity)->exists()) {
+            return redirect()->back()->withErrors(['min_quantity' => 'El descuento ingresado intersecta con otro descuento ya registrado.'])
+                ->withInput($request->input());
+        }
         $product->discounts()->create($request->validated());
 
         return redirect()->route('discounts.index', $product->id)
@@ -56,6 +66,8 @@ class ProductDiscountController extends Controller
      */
     public function show(Product $product, ProductDiscount $productDiscount)
     {
+        $this->authorize('view', ProductDiscount::class);
+
         return view('discounts.show', compact('product', 'productDiscount'));
     }
 
@@ -68,6 +80,8 @@ class ProductDiscountController extends Controller
      */
     public function edit(Product $product, ProductDiscount $productDiscount)
     {
+        $this->authorize('update', ProductDiscount::class);
+
         return view('discounts.edit', compact('product', 'productDiscount'));
     }
 
@@ -81,6 +95,12 @@ class ProductDiscountController extends Controller
      */
     public function update(ProductDiscountRequest $request, Product $product, ProductDiscount $productDiscount)
     {
+        $this->authorize('update', ProductDiscount::class);
+
+        if (ProductDiscount::where('product_id', $product->id)->where('id', '<>', $productDiscount->id)->intersecting($request->min_quantity, $request->max_quantity)->exists()) {
+            return redirect()->back()->withErrors(['min_quantity' => 'El descuento ingresado intersecta con otro descuento ya registrado.'])
+                ->withInput($request->input());
+        }
         $productDiscount->update($request->validated());
 
         return redirect()->route('discounts.index', $product->id)
@@ -96,6 +116,8 @@ class ProductDiscountController extends Controller
      */
     public function destroy(Product $product, ProductDiscount $productDiscount)
     {
+        $this->authorize('delete', ProductDiscount::class);
+
         $productDiscount->delete();
 
         return redirect()->route('discounts.index', $product->id)
