+++ admin_page/app/Http/Controllers/TimeBlockController.php
@@ -14,6 +14,8 @@ class TimeBlockController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', TimeBlock::class);
+
         $timeBlocks = TimeBlock::orderedBlocksPaginated();
         return view('schedule.index', compact('timeBlocks'))
             ->with('rowItem', $this->rowNumber(request()->input('page', 1), TimeBlock::ITEMS_PER_PAGE));
@@ -26,6 +28,8 @@ class TimeBlockController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', TimeBlock::class);
+
         return view('schedule.create');
     }
 
@@ -37,6 +41,12 @@ class TimeBlockController extends Controller
      */
     public function store(TimeBlockRequest $request)
     {
+        $this->authorize('create', TimeBlock::class);
+
+        if (TimeBlock::intersecting($request->start, $request->end)->exists()) {
+            return redirect()->back()->withErrors(['start' => 'El bloque ingresado intersecta con el horario de algún otro bloque.'])
+                ->withInput($request->input());
+        }
         TimeBlock::create($request->validated());
 
         return redirect()->route('schedule.index')
@@ -51,6 +61,8 @@ class TimeBlockController extends Controller
      */
     public function show(TimeBlock $timeBlock)
     {
+        $this->authorize('view', TimeBlock::class);
+
         return view('schedule.show', compact('timeBlock'));
     }
 
@@ -62,6 +74,8 @@ class TimeBlockController extends Controller
      */
     public function edit(TimeBlock $timeBlock)
     {
+        $this->authorize('update', TimeBlock::class);
+
         return view('schedule.edit', compact('timeBlock'));
     }
 
@@ -74,6 +88,12 @@ class TimeBlockController extends Controller
      */
     public function update(TimeBlockRequest $request, TimeBlock $timeBlock)
     {
+        $this->authorize('update', TimeBlock::class);
+
+        if (TimeBlock::where('id', '<>', $timeBlock->id)->intersecting($request->start, $request->end)->exists()) {
+            return redirect()->back()->withErrors(['start' => 'El bloque ingresado intersecta con el horario de algún otro bloque.'])
+                ->withInput($request->input());
+        }
         $timeBlock->update($request->validated());
 
         return redirect()->route('schedule.index')
@@ -88,6 +108,8 @@ class TimeBlockController extends Controller
      */
     public function destroy(TimeBlock $timeBlock)
     {
+        $this->authorize('delete', TimeBlock::class);
+
         $timeBlock->delete();
 
         return redirect()->route('schedule.index')
